pdf/requirements.txt
scikit-learn
fpdf == 1.7.2
pandas == 1.5.1
numpy == 1.24.1
temp == 2020.7.2
db-dtypes == 1.0.5
gcsfs == 2022.11.0
fsspec == 2022.11.0
matplotlib == 3.6.3
google-auth == 2.12.0
google-cloud == 0.34.0
google-crc32c == 1.5.0
python-dateutil == 2.8.2
google-api-core == 2.10.2
google-cloud-core == 2.3.2
functions-framework == 3.2.0
google-auth-oauthlib == 0.5.3
google-cloud-storage == 2.7.0
google-resumable-media == 2.4.0
googleapis-common-protos == 1.56.4
google-cloud-bigquery-storage == 2.16.2